Spider Quotes Scraper 2 v 1.3.1

+++ quotes_scraper/quotes_scraper/quotes_scraper/spiders/quotes_csv.py
@@ -14,12 +14,13 @@ import os
 ## '//li[@class="next"]/a/@href'
 
 class quotesSpider(scrapy.Spider):
-    name = 'quotes_csv'
+    name = 'q_csv'
     start_urls = [
         'http://quotes.toscrape.com/'
     ]
-
-    os.remove("quotes_3.csv")
+    
+    if os.path.exists('quotes_3.csv'):
+        os.remove("quotes_3.csv")
 
     custom_settings = {"FEEDS":{"quotes_3.csv":{"format":"csv"}}}
 
@@ -31,19 +32,20 @@ class quotesSpider(scrapy.Spider):
         if next_page_button_link:
             yield response.follow(next_page_button_link, callback=self.parse_quotes, cb_kwargs={'quotes': quotes})
         else: 
-            yield {
-                'quotes' : quotes,
-                }
+            for i in range(len(quotes)):
+                yield {
+                    'quotes' : quotes,
+                    }
 
     def parse(self, response):
         title = response.xpath('//h1/a/text() ').get()
         quotes = response.xpath('//span[@class="text" and @itemprop="text"]/text()').getall()
         ten_tags = response.xpath('//div[contains(@class, "tags-box")]//span[@class="tag-item"]/a/text()').getall()
         
-        yield {
-            'title': title,
-            'ten_tags': ten_tags
-        }
+        ## yield {
+        ##     'title': title,
+        ##     'ten_tags': ten_tags
+        ## }
 
         next_page_button_link = response.xpath('//li[@class="next"]/a/@href').get()
         if next_page_button_link:
